Sort libraries in Webpack console stats by file size (largest to smallest)
By default, webpack will output a list of all files used in the bundle in order of id (which is based on the order they're used in the source, so not super useful).

Much more useful is to see these files listed in order of file size, from largest to smallest, so you can get a good sense of which libraries are adding the most heft to your bundle. This PR does that.